% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/giraffe.R
\name{get_blobs}
\alias{get_blobs}
\title{Get blobs from the git database}
\usage{
get_blobs(repo = repo, path = "test.csv")
}
\arguments{
\item{repo}{A git2R repo object}

\item{path}{A relative path to the table that you are interested in}
}
\description{
None
}

