% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/giraffe.R
\name{make_tables}
\alias{make_tables}
\title{render diffs of an object in the git database}
\usage{
make_tables(repo = repo, path = "test.csv")
}
\arguments{
\item{repo}{A git2R repo object}

\item{path}{A relative path to the table that you are interested in}
}
\description{
None
}
\examples{
repo <- read_sample_repo()
 make_tables(repo = repo, path = "test.csv")
}

