% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_sample_repo.R
\name{read_sample_repo}
\alias{read_sample_repo}
\title{Read the sample data}
\usage{
read_sample_repo()
}
\arguments{
\item{none}{}
}
\value{
a git repo object
}
\description{
Show the sample data for giraffe
}
\examples{
repo<-read_sample_repo()
 summary(repo)
}

