% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/giraffe.R
\name{render_tables}
\alias{render_tables}
\title{Make a webpage of the tables}
\usage{
render_tables(repo, path)
}
\arguments{
\item{repo}{A git2R repo object}

\item{path}{A relative path to the table that you are interested in}
}
\description{
None
}
\examples{
repo <- read_sample_repo()
 render_tables(repo = repo, path = "test.csv")
}

